+++ client/lib/widget/widget_for_order_dialog/edit_buttons_for_text_container.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class EditButtonsForTextContainer extends StatelessWidget {
+  const EditButtonsForTextContainer({
+    required this.buttons,
+    Key? key,
+  }) : super(key: key);
+  final List<Widget> buttons;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: buttons,
+    );
+  }
+}
